+++ x/evm/types/chain_config.go
@@ -30,6 +30,7 @@ func (cc ChainConfig) EthereumConfig(chainID *big.Int) *params.ChainConfig {
 		MuirGlacierBlock:        getBlockValue(cc.MuirGlacierBlock),
 		BerlinBlock:             getBlockValue(cc.BerlinBlock),
 		LondonBlock:             getBlockValue(cc.LondonBlock),
+		ArrowGlacierBlock:       getBlockValue(cc.ArrowGlacierBlock),
 		TerminalTotalDifficulty: nil,
 		Ethash:                  nil,
 		Clique:                  nil,
@@ -50,6 +51,7 @@ func DefaultChainConfig() ChainConfig {
 	muirGlacierBlock := sdk.ZeroInt()
 	berlinBlock := sdk.ZeroInt()
 	londonBlock := sdk.ZeroInt()
+	arrowGlacierBlock := sdk.ZeroInt()
 
 	return ChainConfig{
 		HomesteadBlock:      &homesteadBlock,
@@ -66,6 +68,7 @@ func DefaultChainConfig() ChainConfig {
 		MuirGlacierBlock:    &muirGlacierBlock,
 		BerlinBlock:         &berlinBlock,
 		LondonBlock:         &londonBlock,
+		ArrowGlacierBlock:   &arrowGlacierBlock,
 	}
 }
 
@@ -118,6 +121,9 @@ func (cc ChainConfig) Validate() error {
 	}
 	if err := validateBlock(cc.LondonBlock); err != nil {
 		return sdkerrors.Wrap(err, "londonBlock")
+	}
+	if err := validateBlock(cc.ArrowGlacierBlock); err != nil {
+		return sdkerrors.Wrap(err, "arrowGlacierBloc")
 	}
 
 	// NOTE: chain ID is not needed to check config order