+++ x/evm/types/chain_config_test.go
@@ -237,6 +237,26 @@ func TestChainConfigValidate(t *testing.T) {
 			true,
 		},
 		{
+			"invalid ArrowGlacierBlock",
+			ChainConfig{
+				HomesteadBlock:      newIntPtr(0),
+				DAOForkBlock:        newIntPtr(0),
+				EIP150Block:         newIntPtr(0),
+				EIP150Hash:          defaultEIP150Hash,
+				EIP155Block:         newIntPtr(0),
+				EIP158Block:         newIntPtr(0),
+				ByzantiumBlock:      newIntPtr(0),
+				ConstantinopleBlock: newIntPtr(0),
+				PetersburgBlock:     newIntPtr(0),
+				IstanbulBlock:       newIntPtr(0),
+				MuirGlacierBlock:    newIntPtr(0),
+				BerlinBlock:         newIntPtr(0),
+				LondonBlock:         newIntPtr(0),
+				ArrowGlacierBlock:   newIntPtr(-1),
+			},
+			true,
+		},
+		{
 			"invalid fork order - skip HomesteadBlock",
 			ChainConfig{
 				DAOForkBlock:        newIntPtr(0),